move function to common directory
