+++ src/schedulers/lab-unstake/lab-unstake.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
 import { ElasticsearchService } from '@nestjs/elasticsearch';
 import { SchedulerRegistry } from '@nestjs/schedule';
-import { SubstrateService } from '@common/index';
+import { strToMilisecond, SubstrateService } from '@common/index';
 import {
   queryLabById,
   retrieveLabUnstakeAmount,
@@ -22,10 +22,10 @@ export class LabUnstakedService implements OnModuleInit {
   ) {}
 
   onModuleInit() {
-    this.timer = this.strToMilisecond(
+    this.timer = strToMilisecond(
       this.gCloudSecretManagerService.getSecret('UNSTAKE_TIMER').toString(),
     );
-    const unstakeInterval: number = this.strToMilisecond(
+    const unstakeInterval: number = strToMilisecond(
       this.gCloudSecretManagerService.getSecret('UNSTAKE_INTERVAL').toString(),
     );
 
@@ -116,28 +116,5 @@ export class LabUnstakedService implements OnModuleInit {
     } finally {
       this.isRunning = false;
     }
-  }
-
-  strToMilisecond(timeFormat: string): number {
-    // time format must DD:HH:MM:SS
-    const splitTimeFormat = timeFormat.split(':');
-
-    const d = Number(splitTimeFormat[0]);
-    const h = Number(splitTimeFormat[1]);
-    const m = Number(splitTimeFormat[2]);
-    const s = Number(splitTimeFormat[3]);
-
-    const dayToMilisecond = d * 24 * 60 * 60 * 1000;
-    const hourToMilisecond = h * 60 * 60 * 1000;
-    const minuteToMilisecond = m * 60 * 1000;
-    const secondToMilisecond = s * 1000;
-
-    const result =
-      dayToMilisecond +
-      hourToMilisecond +
-      minuteToMilisecond +
-      secondToMilisecond;
-
-    return result;
   }
 }
